+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Added
+
+- Modem GPS drawer toggle
+
 ## 1.0.3 - 2025-10-19
 
 ### Fixed